Improve test coverage by ignoring polyfill file
