+++ src/CustomEvent.js
@@ -1,0 +1,12 @@
+// https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent#Polyfill
+export var CustomEvent = (function(c) {
+  if (typeof c === "function") return c
+  function CustomEvent(event, params) {
+    params = params || { bubbles: false, cancelable: false, detail: undefined }
+    var evt = document.createEvent("CustomEvent")
+    evt.initCustomEvent(event, params.bubbles, params.cancelable, params.detail)
+    return evt
+  }
+  CustomEvent.prototype = window.Event.prototype
+  return CustomEvent
+})(window.CustomEvent)
