+++ src/location.js
@@ -1,15 +1,4 @@
-// https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent#Polyfill
-var CustomEvent = (function (c) {
-  if (typeof c === "function") return c
-  function CustomEvent(event, params) {
-    params = params || { bubbles: false, cancelable: false, detail: undefined };
-    var evt = document.createEvent( 'CustomEvent' );
-    evt.initCustomEvent( event, params.bubbles, params.cancelable, params.detail );
-    return evt;
-  }
-  CustomEvent.prototype = window.Event.prototype
-  return CustomEvent
-})(window.CustomEvent)
+import { CustomEvent } from "./CustomEvent"
 
 function wrapHistory(keys) {
   return keys.reduce(function(next, key) {
